backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-forum-rag"
version = "1.0.0"
description = "LangChain Forum RAG Backend with LangServe"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Steve Yu"}
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langserve[all]>=0.3.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "langsmith>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
